final version

+++ src/libhello/lib_calc_trans.h
@@ -1,11 +1,10 @@
 #include <ctype.h>
 #include <math.h>
-#include <stdbool.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
 
 #pragma once
 
-void calculation(double*);
-int translate(char*, double*);
+void calculation(int*);
+int translate(char*, int*);
